convsym neon to a separate file
